Backend/requirements.txt
# Backend Requirements for AI Architect
# Core web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0

# AI and API clients
groq>=0.4.1
requests>=2.31.0
aiohttp>=3.9.0

# Data validation and models
pydantic>=2.5.0

# Image processing
Pillow>=10.0.0

# Database and caching
aiosqlite>=0.19.0

# Web scraping and HTTP requests
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# File uploads
python-multipart>=0.0.6

# Optional: For enhanced functionality
# openai>=1.0.0  # If using OpenAI models
# anthropic>=0.7.0  # If using Claude models

# Development and utilities
typing-extensions>=4.8.0
asyncio  # Built-in with Python 3.7+

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
